requirements.txt
Django>=5.2,<5.3
wagtail>=7.2,<7.3
django-browser-reload>=1.18,<2.0
honcho>=1.1.0,<2.0
python-dotenv>=1.0.1,<2.0
whitenoise>=6.7,<7.0
setuptools>=70,<80
